+++ library/src/main/java/org/mustangproject/EStandard.java
@@ -1,6 +1,6 @@
 package org.mustangproject;
 
 public enum EStandard {
-	facturx, orderx, deliverx, zugferd, cii, ubl
+	facturx, orderx, deliveradvice, zugferd, cii, ubl
 
 }