+++ History.md
@@ -1,3 +1,6 @@
+- upgraded en16931 validation to 1.3.8
+- Be able to embed Cross Industry Delivery Advice
+- update xr to v. 2.2
 2.5.0
 =======
 2022-04-07